split up post hoc global methods for understanding and streamlining
